refactor: rename export_as_terraform to generate_hcl
