+++ test/sandbox/sandbox.go
@@ -427,21 +427,21 @@ func (se StackEntry) ReadGeneratedLocals() []byte {
 	return se.DirEntry.ReadFile(cfg.LocalsFilename)
 }
 
-// ReadGeneratedTerraform will read code that was generated by Terramate for this stack
-// using export_as_terraform blocks.
-// The given name is the name of the export_as_terraform block as indicated by its label.
+// ReadGeneratedHCL will read code that was generated by Terramate for this stack
+// using generate_hcl blocks.
+// The given name is the name of the generate_hcl block as indicated by its label.
 //
 // It will fail the test if there is no generated code available on the stack,
 // since it assumes generated code is expected to be there.
-func (se StackEntry) ReadGeneratedTerraform(name string) string {
+func (se StackEntry) ReadGeneratedHCL(name string) string {
 	se.t.Helper()
 	return string(se.DirEntry.ReadFile(name))
 }
 
-// RemoveGeneratedTerraform will delete the file with generated code from
-// export_as_terraform blocks.
-// The given name is the name of the export_as_terraform block as indicated by its label.
-func (se StackEntry) RemoveGeneratedTerraform(name string) {
+// RemoveGeneratedHCL will delete the file with generated code from
+// generate_hcl blocks.
+// The given name is the name of the generate_hcl block as indicated by its label.
+func (se StackEntry) RemoveGeneratedHCL(name string) {
 	se.t.Helper()
 	se.DirEntry.RemoveFile(name)
 }
